AI_Friends/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.37.0
pydantic>=2.7.0
python-dotenv>=1.0.1
Jinja2>=3.1.4
